componentize import
